CNPM/requirements.txt
streamlit
pandas
Pillow
plotly